put enum and class to separate file
