+++ WalletWasabi.Fluent/Model/TimePeriodOption.cs
@@ -1,0 +1,21 @@
+using WalletWasabi.Models;
+
+namespace WalletWasabi.Fluent.Model
+{
+	public enum TimePeriodOption
+	{
+		All,
+        [FriendlyName("1D")]
+        Day,
+        [FriendlyName("1W")]
+        Week,
+        [FriendlyName("1M")]
+        Month,
+        [FriendlyName("3M")]
+        ThreeMonths,
+        [FriendlyName("6M")]
+        SixMonths,
+        [FriendlyName("1Y")]
+        Year
+	}
+}
